+++ pyam/utils.py
@@ -20,11 +20,15 @@ except ImportError:
 logger = logging.getLogger(__name__)
 
 # common indices
+DEFAULT_META_INDEX = ['model', 'scenario']
 META_IDX = ['model', 'scenario']
 YEAR_IDX = ['model', 'scenario', 'region', 'year']
 IAMC_IDX = ['model', 'scenario', 'region', 'variable', 'unit']
 SORT_IDX = ['model', 'scenario', 'variable', 'year', 'region']
 LONG_IDX = IAMC_IDX + ['year']
+
+# required columns
+REQUIRED_COLS = ['region', 'variable', 'unit']
 
 # illegal terms for data/meta column names to prevent attribute conflicts
 ILLEGAL_COLS = ['data', 'meta']
@@ -126,14 +130,14 @@ def read_pandas(path, default_sheet='data', *args, **kwargs):
 
 def read_file(path, *args, **kwargs):
     """Read data from a file"""
-    format_kwargs = {}
     # extract kwargs that are intended for `format_data`
+    format_kwargs = dict(index=kwargs.pop('index'))
     for c in [i for i in IAMC_IDX + ['year', 'time', 'value'] if i in kwargs]:
         format_kwargs[c] = kwargs.pop(c)
     return format_data(read_pandas(path, *args, **kwargs), **format_kwargs)
 
 
-def format_data(df, **kwargs):
+def format_data(df, index, **kwargs):
     """Convert a pandas.Dataframe or pandas.Series to the required format"""
     if isinstance(df, pd.Series):
         df.name = df.name or 'value'
@@ -214,29 +218,33 @@ def format_data(df, **kwargs):
     if conflict_cols:
         msg = f'Column name {conflict_cols} is illegal for timeseries data.\n'
         _args = ', '.join([f"{i}_1='{i}'" for i in conflict_cols])
-        msg += f'Use `IamDataFrame(..., {_args})` to rename at initalization.'
+        msg += f'Use `IamDataFrame(..., {_args})` to rename at initialization.'
         raise ValueError(msg)
 
-    # format columns to lower-case and check that all required columns exist
-    if not set(IAMC_IDX).issubset(set(df.columns)):
-        missing = list(set(IAMC_IDX) - set(df.columns))
-        raise ValueError("missing required columns `{}`!".format(missing))
+    # check that index and required columns exist
+    missing_index = [c for c in index if c not in df.columns]
+    if missing_index:
+        raise ValueError(f'Missing index columns `{missing_index}`!')
+
+    missing_required_col = [c for c in REQUIRED_COLS if c not in df.columns]
+    if missing_required_col:
+        raise ValueError(f'Missing required columns `{missing_required_col}`!')
 
     # check whether data in wide format (IAMC) or long format (`value` column)
     if 'value' in df.columns:
         # check if time column is given as `year` (int) or `time` (datetime)
-        cols = df.columns
-        if 'year' in cols:
+        if 'year' in df.columns:
             time_col = 'year'
-        elif 'time' in cols:
+        elif 'time' in df.columns:
             time_col = 'time'
         else:
-            msg = 'invalid time format, must have either `year` or `time`!'
+            msg = 'Invalid time format, must have either `year` or `time`!'
             raise ValueError(msg)
-        extra_cols = list(set(cols) - set(IAMC_IDX + [time_col, 'value']))
+        extra_cols = [c for c in df.columns
+                      if c not in index + REQUIRED_COLS + [time_col, 'value']]
     else:
         # if in wide format, check if columns are years (int) or datetime
-        cols = set(df.columns) - set(IAMC_IDX)
+        cols = [c for c in df.columns if c not in index + REQUIRED_COLS]
         year_cols, time_cols, extra_cols = [], [], []
         for i in cols:
             try:
@@ -257,7 +265,8 @@ def format_data(df, **kwargs):
         else:
             msg = 'invalid column format, must be either years or `datetime`!'
             raise ValueError(msg)
-        df = pd.melt(df, id_vars=IAMC_IDX + extra_cols, var_name=time_col,
+        cols = index + REQUIRED_COLS + extra_cols
+        df = pd.melt(df, id_vars=cols, var_name=time_col,
                      value_vars=sorted(melt_cols), value_name='value')
 
     # cast value column to numeric and drop nan
@@ -273,7 +282,7 @@ def format_data(df, **kwargs):
         _raise_data_error('empty cells in `data`', df.loc[null_rows])
 
     # check for duplicates and empty data
-    idx_cols = IAMC_IDX + [time_col] + extra_cols
+    idx_cols = index + REQUIRED_COLS + [time_col] + extra_cols
     rows = df[idx_cols].duplicated()
     if any(rows):
         _raise_data_error('duplicate rows in `data`', df.loc[rows, idx_cols])
@@ -282,7 +291,7 @@ def format_data(df, **kwargs):
         logger.warning('Formatted data is empty!')
 
     df = format_time_col(sort_data(df, idx_cols), time_col)
-    return df, time_col, extra_cols
+    return df.set_index(idx_cols).value, index, time_col, extra_cols
 
 
 def format_time_col(data, time_col):